I submitted this on time.
listen
LISTEN
I forgot to save, it happens to the best, and I'm not the best, so it's only reasonable it will happen to me.
I can tell you are not listening.
Did I really submit after the deadline? What even is time? According to special relativity, time and space are one fabric, and are directly tied to each other. Suppose I was moving through space-time in the speed of light, c, Since my four-momentum is always equal to c, my "speed" through time will be equivalent to 0. therefore, The deadline couldn't have passed for me. For you, it might have passed, but time is relative. Therefore, I'm submitting this on time. If you are Shalev, you might be thinking "b... bu... bu... but how c... come you... you... are back to this tim... time?" First, Shalev, you need to relax. This is physics, which is a real science, unlike biology. go back to Kazakh to swim with your amphibian friends. I'm back in this time because I created and passed through an Einstein Rosen Bridge, commonly known as a wormhole. Usually, when speaking of wormholes, what comes to mind is spatial displacement, but most forget temporal displacement is possible as well (as mentioned before, time and space are equivalent). So, now  that Shalev is satisfied, we can all agree I HAVE SUBMITTED THIS PROJECT ON TIME.
Thank you for listening to my TED talk.